app/requirements.txt
ultralytics==8.1.22
tensorflow==2.15.0
keras==2.15.0
fastapi==0.110.1
uvicorn==0.29.0
weaviate-client==4.5.5
transformers==4.39.3
openai==1.21.2
manga-ocr==0.1.11
natsort==8.4.0
google-api-core==2.18.0
google-cloud-storage==2.16.0
celery==5.4.0
python-multipart==0.0.9
redis==5.0.3
flower==2.0.1
python-dotenv==1.0.1
Pygments==2.17.2
jieba==0.42.1